Moved some tests to *derive_lib crate,added some non-compiling #[sabi_trait]traits.
