+++ abi_stable/Cargo.toml
@@ -45,7 +45,6 @@ generational-arena = "0.2"
 crossbeam-channel = {version="0.3",optional=true}
 
 [dev-dependencies]
-abi_stable_derive_lib= {version="0.4",path="../abi_stable_derive_lib"}
 bincode = "1.1"
 crossbeam-utils = "0.6"
 serde_json = "1.0"
