+++ abi_stable/src/trait_object_test.rs
@@ -1,8 +1,14 @@
 use std::fmt::Debug;
-use crate::*;
+use crate::{
+    *,
+    std_types::{RBox,RArc},
+    sabi_trait::prelude::*,
+};
+
+//////////////////////////////////////
 
 #[sabi_trait]
-//#[sabi(debug_print_trait)]
+#[sabi(debug_print_trait)]
 pub trait RSomething<T>:Send+Sync+Clone+Debug{
     type Element:Debug;
 
@@ -11,39 +17,85 @@ pub trait RSomething<T>:Send+Sync+Clone+Debug{
     fn get_mut(&mut self)->&mut Self::Element;
 
     #[sabi(last_prefix_field)]
-    fn into_value(self)->Self::Element
-    where Self:Sized;
+    fn into_value(self)->Self::Element;
 }
 
-///////////////////////////////////////////////////////////////////////////
-///////                Generated Code                               ///////
-///////////////////////////////////////////////////////////////////////////
+impl RSomething<()> for u32{
+    type Element=u32;
+
+    fn get(&self)->&Self::Element{
+        self
+    }
+    fn get_mut(&mut self)->&mut Self::Element{
+        self
+    }
+    fn into_value(self)->Self::Element{
+        self
+    }
+}
+
+//////////////////////////////////////
+
+#[sabi_trait]
+//#[sabi(debug_print_trait)]
+pub trait RSomethingElse<T:Copy>:Send+Sync+Clone+Debug{
+    fn get(&self)->&T;
+    
+    fn get_mut(&mut self)->&mut T;
+
+    #[sabi(last_prefix_field)]
+    fn into_value(self)->T;
+}
+
+impl RSomethingElse<u32> for u32{
+    fn get(&self)->&u32{
+        self
+    }
+    fn get_mut(&mut self)->&mut u32{
+        self
+    }
+    fn into_value(self)->u32{
+        self
+    }
+}
+
+//////////////////////////////////////
+
+#[sabi_trait]
+#[sabi(debug_print_trait)]
+pub trait RFoo<'a,T:Copy+'a>:Send+Sync+Clone+Debug{
+    fn get(&'a self)->&'a T;
+}
+
+//////////////////////////////////////
+fn _rsomething_assertions(
+    trait_0:&mut dyn RSomething<(),Element=()>,
+    trait_1:&mut dyn RSomethingElse<u32>,
+)where
+    RSomething_TO<'static,RBox<()>,(),()>:RSomething<()>+Send+Sync+Clone+Debug,
+    RSomethingElse_TO<'static,RBox<()>,u32>:RSomethingElse<u32>+Send+Sync+Clone+Debug,
+{
+    trait_0.get();
+    trait_0.get_mut();
+    trait_1.get();
+    trait_1.get_mut();
+
+    let _object_0=RSomething_from_value::<_,YesImplAny,()>(100_u32);
+    let _erased_0=RSomething_from_ptr::<_,NoImplAny,()>(RBox::new(100_u32));
+
+    let _object_1:RSomethingElse_TO<RBox<()>,u32>=
+        RSomethingElse_from_value::<_,YesImplAny,u32>(100_u32);
+
+    let _erased_1:RSomethingElse_TO<RBox<()>,u32>=
+        RSomethingElse_from_ptr::<_,NoImplAny,u32>(RBox::new(100_u32));
+
+}
+//////////////////////////////////////
 
 
 #[cfg(test)]
 mod tests{
     use super::*;
-
-    use crate::{
-        std_types::RBox,
-        sabi_trait::prelude::*,
-    };
-
-    impl RSomething<()> for u32{
-        type Element=u32;
-
-        fn get(&self)->&Self::Element{
-            self
-        }
-        fn get_mut(&mut self)->&mut Self::Element{
-            self
-        }
-        fn into_value(self)->Self::Element
-        where Self:Sized
-        {
-            self
-        }
-    }
 
     fn assert_sync<T:Sync>(_:&T){}
     fn assert_debug<T:Debug>(_:&T){}
