+++ abi_stable_derive_lib/src/utils.rs
@@ -1,8 +1,11 @@
 use std::time::Instant;
 
+use abi_stable_shared::test_utils::{FileSpan};
+
 use core_extensions::measure_time::MyDuration;
 use quote::{ToTokens};
 use proc_macro2::TokenStream as TokenStream2;
+
 
 #[derive(Debug,Copy,Clone,PartialEq,Eq,Hash)]
 pub struct NoTokens;
@@ -10,25 +13,6 @@ pub struct NoTokens;
 impl ToTokens for NoTokens {
     fn to_tokens(&self, _: &mut TokenStream2) {}
 }
-
-
-
-
-#[derive(Debug, Clone,Copy)]
-pub struct FileSpan {
-    pub file: &'static str,
-    pub line: u32,
-}
-macro_rules! file_span {
-    () => {{
-        use crate::utils::FileSpan;
-        FileSpan {
-            file: file!(),
-            line: line!(),
-        }
-    }};
-}
-
 
 
 pub struct PrintDurationOnDrop{
